Support for shims.
The approach is the simplest I could imagine. I assume there must be better
ways of doing it.

For declaring shims, you provide an array in the configuration:
```js
compileModules(inputTree, {
  formatter: 'bundle',
  shims: [
    'Ember',
    'moment'
  ]
})
```

For each entry in that array, a temporary file will be created with the
downcased name of the shim, exporting the global.

By example, the former config will create a temp file named `ember.js`
containing the following code:

```js
export default Ember;
```

That way the module compiler will work.

The downside of this approach is that the resultant build will contain a faily
useless line exporting a global. A potential workaround could be delete that
useless line using a String.replace.

No tests yet. I would like some feedback first.