Improve structure of cache dir
This improves the structure of the cache dir used by this package by
keeping one dir and changing the files in it, rather than creating a
new cache sub-dir for every transpiler run.

Having a single cache dir allows files to be linked from the cache dir
to the dest dir on each rebuild, while maintaining the structure of the
src dir.